proxy/src/tds_handler.py
import os
import socket
import logging
from connection_pool import CursorAwareConnectionPool
from prepared_statements import PreparedStatementManager
from cursor_manager import CursorManager

class TDSHandler:
    def __init__(self, sock):
        self.sock = sock
        self.pool = CursorAwareConnectionPool(
            minconn=5,
            maxconn=20,
            host=os.getenv("PG_HOST"),
            database=os.getenv("PG_DB"),
            user=os.getenv("PG_USER"),
            password=os.getenv("PG_PASSWORD")
        )
        self.statement_mgr = PreparedStatementManager()
        self.cursor_mgr = CursorManager()

    def handle_client(self):
        try:
            while True:
                packet = self.sock.recv(4096)
                if not packet:
                    break
                
                # Handle TDS protocol
                if packet[0] == 0x03:  # SQL Batch
                    self.handle_sql_batch(packet)
                elif packet[0] == 0x04:  # RPC
                    self.handle_rpc(packet)
                
        finally:
            self.sock.close()

    def handle_rpc(self, packet):
        # Handle stored procedures and prepared statements
        pass
